+++ CHANGELOG.rst
@@ -1,9 +1,18 @@
-**TBD - TBD/TBD/TBD**
+**3.2.8 - 12/16/24**
 
-  - Update how addtional seed is defined in RandomnessManager
+  - Feature: Update how addtional seed is defined in RandomnessManager
 
-**3.2.5 - TBD/TBD/TBD
+**3.2.7 - 12/12/24**
 
+  - Type-hinting: Fix mypy errors in vivarium/framework/state_machine.py
+
+**3.2.6 - 12/12/24**
+
+  - Change Jenkins conda env name
+
+**3.2.5 - 12/11/24**
+
+  - Type-hinting: Fix mypy errors in vivarium/framework/results/interface.py
   - Type-hinting: Fix mypy errors in vivarium/component.py
   - Type-hinting: Fix mypy errors in vivarium/framework/results/observer.py